AI: Use memory block in RTC packet.
