+++ trunk/src/kernel/srs_kernel_flv.cpp
@@ -263,60 +263,6 @@ void SrsMessageHeader::initialize_video(int size, uint32_t time, int stream)
     prefer_cid = RTMP_CID_Video;
 }
 
-SrsMemoryBlock::SrsMemoryBlock()
-{
-    payload_ = NULL;
-    size_ = 0;
-}
-
-SrsMemoryBlock::~SrsMemoryBlock()
-{
-    srs_freepa(payload_);
-}
-
-void SrsMemoryBlock::create(int size)
-{
-    srs_assert(size >= 0);
-
-    // Free existing payload
-    srs_freepa(payload_);
-
-    // Allocate new buffer
-    if (size > 0) {
-        payload_ = new char[size];
-        memset(payload_, 0, size);
-        size_ = size;
-    } else {
-        payload_ = NULL;
-        size_ = 0;
-    }
-}
-
-void SrsMemoryBlock::create(char *data, int size)
-{
-    srs_assert(size >= 0);
-
-    create(size);
-
-    // Copy data if provided
-    if (data && size > 0) {
-        memcpy(payload_, data, size);
-        size_ = size;
-    }
-}
-
-void SrsMemoryBlock::attach(char *data, int size)
-{
-    srs_assert(size >= 0);
-
-    // Free existing payload
-    srs_freepa(payload_);
-
-    // Attach new buffer
-    payload_ = data;
-    size_ = size;
-}
-
 SrsCommonMessage::SrsCommonMessage()
 {
     payload_ = SrsSharedPtr<SrsMemoryBlock>(NULL);
