+++ trunk/src/kernel/srs_kernel_rtc_rtp.cpp
@@ -761,7 +761,7 @@ SrsRtpPacket::SrsRtpPacket()
 {
     payload_ = NULL;
     payload_type_ = SrsRtpPacketPayloadTypeUnknown;
-    shared_buffer_ = NULL;
+    shared_buffer_ = SrsSharedPtr<SrsMemoryBlock>(NULL);
     actual_buffer_size_ = 0;
 
     nalu_type = 0;
@@ -776,7 +776,7 @@ SrsRtpPacket::SrsRtpPacket()
 SrsRtpPacket::~SrsRtpPacket()
 {
     srs_freep(payload_);
-    srs_freep(shared_buffer_);
+    // shared_buffer_ automatically cleaned up by SrsSharedPtr
 }
 
 char *SrsRtpPacket::wrap(int size)
@@ -785,19 +785,18 @@ char *SrsRtpPacket::wrap(int size)
     actual_buffer_size_ = size;
 
     // If the buffer is large enough, reuse it.
-    if (shared_buffer_ && shared_buffer_->size() >= size) {
+    if (shared_buffer_.get() && shared_buffer_->size() >= size) {
         return shared_buffer_->payload();
     }
 
-    // Create a large enough message, with under-layer buffer.
-    srs_freep(shared_buffer_);
-    shared_buffer_ = new SrsSharedPtrMessage();
+    // Create a large enough memory block, with under-layer buffer.
+    shared_buffer_ = SrsSharedPtr<SrsMemoryBlock>(new SrsMemoryBlock());
 
-    // Create under-layer buffer for new message
+    // Create under-layer buffer for new memory block
     // For RTC, we use larger under-layer buffer for each packet.
     int nb_buffer = srs_max(size, kRtpPacketSize);
     char *buf = new char[nb_buffer];
-    shared_buffer_->wrap(buf, nb_buffer);
+    shared_buffer_->attach(buf, nb_buffer);
 
     ++_srs_pps_objs_rbuf->sugar;
 
@@ -811,18 +810,14 @@ char *SrsRtpPacket::wrap(char *data, int size)
     return buf;
 }
 
-char *SrsRtpPacket::wrap(SrsSharedPtrMessage *msg)
+char *SrsRtpPacket::wrap(SrsSharedPtr<SrsMemoryBlock> block)
 {
-    // Generally, the wrap(msg) is used for RTMP to RTC, where the msg
-    // is not generated by RTC.
-    srs_freep(shared_buffer_);
+    // Copy the shared memory block.
+    shared_buffer_ = block;
+    // If we wrap a memory block, the size of packet equals to the block size.
+    actual_buffer_size_ = shared_buffer_.get() ? shared_buffer_->size() : 0;
 
-    // Copy from the new message.
-    shared_buffer_ = msg->copy();
-    // If we wrap a message, the size of packet equals to the message size.
-    actual_buffer_size_ = shared_buffer_->size();
-
-    return msg->payload();
+    return shared_buffer_.get() ? shared_buffer_->payload() : NULL;
 }
 
 SrsRtpPacket *SrsRtpPacket::copy()
@@ -834,7 +829,7 @@ SrsRtpPacket *SrsRtpPacket::copy()
     cp->payload_type_ = payload_type_;
 
     cp->nalu_type = nalu_type;
-    cp->shared_buffer_ = shared_buffer_ ? shared_buffer_->copy2() : NULL;
+    cp->shared_buffer_ = shared_buffer_;  // Copy shared pointer
     cp->actual_buffer_size_ = actual_buffer_size_;
     cp->frame_type = frame_type;
